Fix some warnings

+++ src/core/cc_fileformat.h
@@ -466,7 +466,7 @@ namespace Parser {
     {
         std::vector<std::tuple<uint32_t, Iter, size_t> > result;
         while (begin != end) {
-            auto length = (uint32_t)std::distance(begin, end);
+            auto length = std::distance(begin, end);
             if (length < 8) {
                 return result;
             }
@@ -475,7 +475,7 @@ namespace Parser {
             auto size = get_big_long(begin);
             begin += 4;
             length = std::distance(begin, end);
-            if (length < size + 8) {
+            if (static_cast<uint32_t>(length) < size + 8) {
                 return result;
             }
             auto data = begin;
